fixed openmp issus with VC2013 as in ticket #80

+++ trunk/Shark/include/shark/Unsupervised/RBM/GradientApproximations/ContrastiveDivergence.h
@@ -148,7 +148,7 @@ public:
 		std::size_t numBatches = batchesForTraining/threads;
 		
 		
-		SHARK_PARALLEL_FOR(size_t t = 0; t < threads; ++t){
+		SHARK_PARALLEL_FOR(int t = 0; t < (int)threads; ++t){
 			AverageEnergyGradient<RBM> empiricalAverage(mpe_rbm);
 			AverageEnergyGradient<RBM> modelAverage(mpe_rbm);
